requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pypdf
langchain
langchain-community
google-generativeai